[Feat(SignInActivity.kt)] 코드 전면 수정
1. 자동로그인 로직 전면 삭제
